Materials for OLCF Frontier Workshop
Add in all the files necessary to run a simple example of using
SmartSim to run the MOM6 ocean model with a CNN embedded inside of
the numerical solver. The main driver is in
smartsim_drivers/double_gyre/call_MOM6.py

This example starts up a colocated database that runs alongside
the simulation. Users can modify how many resources are dedicated
to the database or the simulation. Additionally, by passing
the --enable-pinning flag to the call_MOM6.py driver, users can
see the difference in performace when both the applicaton and
the database are pinned to specific processors